Added two items; introduce new structure
Added Cold Denial and The Last Dance. Started a new organization structure; list out wishlist items and group by max infusion cap; will make it easier to add new items to bottom of the list as we move forward and not have to locate them in specific weapon type groupings.