More granular testing
Discovered that DIM will match only the first match it finds from a wishlist. Subsequent wish list items will be ignored.

For more granular wishlist prep, the more specific a combo is, the higher in the list it should be. More general/broad (say, only wishlisting final two columns) can be done at the end of a list.